requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
db-dtypes==1.4.2
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.0
gspread-dataframe==4.0.0
idna==3.10
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==77.0.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
python-dotenv==1.0.0
